# api-docs.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "Uwazi API Documentation",
                 "version":  "1.0.0",
                 "description":  "API documentation for the Uwazi Social Justice Tracking Application"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5000"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "AdminAuth":  {
                                                                 "type":  "apiKey",
                                                                 "in":  "header",
                                                                 "name":  "X-User-Id",
                                                                 "description":  "Admin user ID used for protected admin endpoints"
                                                             }
                                           }
                   },
    "paths":  {
...
                  "/api/users":  {
                                     "post":  {
                                                  "summary":  "Create a new user",
                                                  "tags":  [
                                                               "Users"
                                                           ],
                                                  "requestBody":  {
                                                                      "required":  true,
                                                                      "content":  {
                                                                                      "application/json":  {
                                                                                                               "schema":  {
                                                                                                                              "type":  "object",
                                                                                                                              "required":  [
                                                                                                                                               "email",
                                                                                                                                               "name"
                                                                                                                                           ],
                                                                                                                              "properties":  {
                                                                                                                                                 "email":  {
                                                                                                                                                               "type":  "string",
                                                                                                                                                               "format":  "email"
                                                                                                                                                           },
                                                                                                                                                 "name":  {
                                                                                                                                                              "type":  "string"
                                                                                                                                                          },
                                                                                                                                                 "role":  {
                                                                                                                                                              "type":  "string",
                                                                                                                                                              "enum":  [
                                                                                                                                                                           "admin",
                                                                                                                                                                           "reviewer",
                                                                                                                                                                           "viewer"
                                                                                                                                                                       ],
                                                                                                                                                              "default":  "viewer"
                                                                                                                                                          }
                                                                                                                                             }
                                                                                                                          }
                                                                                                           }
                                                                                  }
                                                                  },
                                                  "responses":  {
                                                                    "201":  {
                                                                                "description":  "User created successfully"
                                                                            },
                                                                    "400":  {
                                                                                "description":  "Validation error"
                                                                            },
                                                                    "409":  {
                                                                                "description":  "User with this email already exists"
                                                                            }
                                                                }
                                              },
                                     "get":  {
                                                 "summary":  "Get all users",
                                                 "tags":  [
                                                              "Users"
                                                          ],
                                                 "parameters":  [
                                                                    {
                                                                        "in":  "query",
                                                                        "name":  "role",
                                                                        "schema":  {
                                                                                       "type":  "string",
                                                                                       "enum":  [
                                                                                                    "admin",
                                                                                                    "reviewer",
                                                                                                    "viewer"
                                                                                                ]
                                                                                   },
                                                                        "description":  "Optional role filter"
                                                                    },
                                                                    {
                                                                        "in":  "query",
                                                                        "name":  "limit",
                                                                        "schema":  {
                                                                                       "type":  "integer",
                                                                                       "default":  10
                                                                                   }
                                                                    },
                                                                    {
                                                                        "in":  "query",
                                                                        "name":  "skip",
                                                                        "schema":  {
                                                                                       "type":  "integer",
                                                                                       "default":  0
                                                                                   }
                                                                    }
                                                                ],
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Paginated list of users"
                                                                           }
                                                               }
                                             }
                                 },
                  "/api/users/{id}":  {
                                          "get":  {
                                                      "summary":  "Get a single user by ID",
                                                      "tags":  [
                                                                   "Users"
                                                               ],
                                                      "parameters":  [
                                                                         {
                                                                             "in":  "path",
                                                                             "name":  "id",
                                                                             "required":  true,
                                                                             "schema":  {
                                                                                            "type":  "string"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "User details"
                                                                                },
                                                                        "404":  {
                                                                                    "description":  "User not found"
                                                                                }
                                                                    }
                                                  },
                                          "put":  {
                                                      "summary":  "Update a user",
                                                      "tags":  [
                                                                   "Users"
                                                               ],
                                                      "parameters":  [
                                                                         {
                                                                             "in":  "path",
                                                                             "name":  "id",
                                                                             "required":  true,
                                                                             "schema":  {
                                                                                            "type":  "string"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "requestBody":  {
                                                                          "required":  true,
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "type":  "object",
                                                                                                                                  "properties":  {
                                                                                                                                                     "email":  {
                                                                                                                                                                   "type":  "string",
                                                                                                                                                                   "format":  "email"
                                                                                                                                                               },
                                                                                                                                                     "name":  {
                                                                                                                                                                  "type":  "string"
                                                                                                                                                              }
                                                                                                                                                 }
                                                                                                                              }
                                                                                                               }
                                                                                      }
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "User updated successfully"
                                                                                },
                                                                        "400":  {
                                                                                    "description":  "Validation error"
                                                                                },
                                                                        "404":  {
                                                                                    "description":  "User not found"
                                                                                }
                                                                    }
                                                  },
                                          "delete":  {
                                                         "summary":  "Delete a user",
                                                         "tags":  [
                                                                      "Users"
                                                                  ],
                                                         "parameters":  [
                                                                            {
                                                                                "in":  "path",
                                                                                "name":  "id",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "User deleted successfully"
                                                                                   },
                                                                           "404":  {
                                                                                       "description":  "User not found"
                                                                                   }
                                                                       }
                                                     }
                                      },
                  "/api/users/{id}/role":  {
                                               "patch":  {
                                                             "summary":  "Update user role",
                                                             "tags":  [
                                                                          "Users"
                                                                      ],
                                                             "parameters":  [
                                                                                {
                                                                                    "in":  "path",
                                                                                    "name":  "id",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "requestBody":  {
                                                                                 "required":  true,
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "type":  "object",
                                                                                                                                         "required":  [
                                                                                                                                                          "role"
                                                                                                                                                      ],
                                                                                                                                         "properties":  {
                                                                                                                                                            "role":  {
                                                                                                                                                                         "type":  "string",
                                                                                                                                                                         "enum":  [
                                                                                                                                                                                      "admin",
                                                                                                                                                                                      "reviewer",
                                                                                                                                                                                      "viewer"
                                                                                                                                                                                  ]
                                                                                                                                                                     }
                                                                                                                                                        }
                                                                                                                                     }
                                                                                                                      }
                                                                                             }
                                                                             },
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "User role updated"
                                                                                       },
                                                                               "400":  {
                                                                                           "description":  "Invalid role"
                                                                                       },
                                                                               "404":  {
                                                                                           "description":  "User not found"
                                                                                       }
                                                                           }
                                                         }
                                           }
              },
...
}